requirements.txt
asgiref==3.3.1
confusable-homoglyphs==3.2.0
Django==3.1.5
django-bootstrap3==14.2.0
django-registration==2.4.1
django-tinymce==3.2.0
djangorestframework==3.12.2
importlib-metadata==1.7.0
Pillow==8.0.1
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
zipp==3.4.0
